add avto
