+++ src/components/About.js
@@ -2,8 +2,20 @@
 
 export default function About() {
     return (
+			
         <div>
-            Hello world!
-        </div>
+          <h1>25 Awesome About Us Page Examples</h1>	
+					<p>The design, written content, and visual or video elements 
+					together tell an important story 
+					about who you are and why you do it.
+
+					How can you make the most of this integral part of your marketing strategy?
+
+					In this article, you’ll learn what makes an exceptional About Us page and 
+					find 25 about us page examples of the best ones out there 
+					to inspire your own About Us page design and content.</p>
+				</div>
+		
+				
     )
 }
