+++ src/components/Contacts.js
@@ -1,9 +1,33 @@
-export default function Contacts() {
-    return (
-        <div>
-            <div>Телефон: 847829374</div>
-            <div>Телеграм: @fy14geektech</div>
-            <div>Email: fy14geektech@gmail.com</div>
-        </div>
-    )
+import ListGroup from 'react-bootstrap/ListGroup';
+
+function DefaultExample() {
+  return (
+    <ListGroup>
+      <ListGroup.Item>Телефон: +996 (776 44-46-55)</ListGroup.Item>
+      <ListGroup.Item>Email- @fy14geektech</ListGroup.Item>
+      <ListGroup.Item>Телеграмм: fy14geektech@gmail</ListGroup.Item>
+      <ListGroup.Item>Адрес: ул Ибримова 139</ListGroup.Item>
+    </ListGroup>
+  );
 }
+
+export default DefaultExample;
+
+
+
+
+
+
+
+
+
+
+// export default function Contacts() {
+//     return (
+//         <div>
+//             <div>Телефон: 847829374</div>
+//             <div>Телеграм: @fy14geektech</div>
+//             <div>Email: fy14geektech@gmail.com</div>
+//         </div>
+//     )
+// }
