+++ src/components/Main/Nav.js
@@ -10,7 +10,6 @@ export default function Nav() {
             <Link to="/about">О нас</Link>&nbsp;&nbsp;&nbsp;
             <Link to="/login">Авторизация</Link>&nbsp;&nbsp;&nbsp;
             <Link to="/contacts">Контакты</Link>&nbsp;&nbsp;&nbsp;
-						<Link to="/password">Пароль</Link>&nbsp;&nbsp;&nbsp;
             <Link to="/lorem">Lorem</Link>
           </div>
           <BasketList/>
